requirements.txt
google-cloud==0.34.0
google-cloud-datastore==1.7.0
google-cloud-pubsub==0.35.4